+++ packages/engine/Source/Shaders/GlobeFS.glsl
@@ -577,6 +577,7 @@ void main()
 #endif
 
 #ifdef HAS_VECTOR_LAYER
+    finalColor = vectorPolygonRender(v_textureCoordinates.xy, finalColor);
     finalColor = vectorPolylineRender(v_textureCoordinates.xy, finalColor);
 #endif